+++ src/main/java/com/amagari/translationtool/mixin/client/MinecraftMixin.java
@@ -1,5 +1,6 @@
 package com.amagari.translationtool.mixin.client;
 
+import com.amagari.translationtool.client.bilingual.BilingualLanguageController;
 import com.amagari.translationtool.client.WorldLanguageClient;
 import com.amagari.translationtool.client.WorldLanguageContext;
 import com.amagari.translationtool.client.paratranz.ParaTranzContext;
@@ -35,6 +36,7 @@ public class MinecraftMixin {
 	@Inject(method = "disconnect(Lnet/minecraft/client/gui/screens/Screen;ZZ)V", at = @At("HEAD"))
 	private void amagari_translation_tool$clearWorldLanguageDirectory(Screen screen, boolean transferring, boolean clearDownloadedPacks, CallbackInfo callbackInfo) {
 		WorldLanguageContext.leaveWorld();
+		BilingualLanguageController.resetSessionState((Minecraft) (Object) this);
 		ParaTranzContext.resetSessionState();
 	}
 }